pyproject.toml
# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "source"}

[project]
name = "solution"
version = "1.1.4"
description = "Data transfer from Amazon S3 Glacier vaults to Amazon S3"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache Software License"}
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Framework :: AWS CDK :: 2",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]
dependencies = [
    "boto3==1.34.91",
    "aws-cdk-lib==2.170.0",
    "aws-cdk.aws-glue-alpha==2.170.0a0",
    "aws-cdk.aws-servicecatalogappregistry-alpha==2.170.0a0",
    "constructs>=10.0.0,<11.0.0",
    "cdk-nag==2.28.12",
    "defusedxml>=0.7.1",
]

[project.scripts]
cdk-app = "solution.app:main"
mock-glacier = "solution.application.mocking.mock_glacier_generator:write_mock_glacier_data"

[project.optional-dependencies]
dev = [
    "tox==4.14.2",
    "black==23.12.1",
    "pytest==7.4.3",
    "pytest-cov==5.0.0",
    "cdk-nag==2.28.12",
    "mypy==1.7.1",
    "moto==4.1.15",
    "pyspark==3.5.1",
    "types-pyyaml==6.0.12.20240311",
    "aws-lambda-powertools[aws-sdk]==2.37.0",
    "boto3-stubs-lite[essential]==1.34.91",
    "boto3-stubs-lite[cloudformation]==1.34.91",
    "boto3-stubs-lite[dynamodb]==1.34.91",
    "boto3-stubs-lite[sqs]==1.34.91",
    "boto3-stubs-lite[sns]==1.34.91",
    "boto3-stubs-lite[s3]==1.34.91",
    "boto3-stubs-lite[iam]==1.34.91",
    "boto3-stubs-lite[stepfunctions]==1.34.91",
    "boto3-stubs-lite[glacier]==1.34.91",
    "boto3-stubs-lite[events]==1.34.91",
    "boto3-stubs-lite[ssm]==1.34.91",
    "boto3-stubs[logs]==1.34.91",
    "importlib-resources>=5.0"
]

[tool.isort]
profile = "black"
known_first_party = "solution"

[tool.bandit]
exclude_dirs = ["tests"]